clippy: just block

+++ pallets/subtensor/tests/block_step.rs
@@ -871,7 +871,7 @@ fn test_emission_based_on_registration_status() {
 
         let block: u64 = 0;
         // drain the emission tuples for the subnet with registration on
-        SubtensorModule::drain_emission(block as u64);
+        SubtensorModule::drain_emission(block);
         // Turn on registration for the subnet with registration off
         SubtensorModule::set_network_registration_allowed(netuid_off, true);
         SubtensorModule::set_network_registration_allowed(netuid_on, false);
